requirements.txt
aiogram~=2.19
python-dotenv~=0.20.0
requests~=2.27.1
bs4~=0.0.1
beautifulsoup4~=4.10.0